migrated slideshow to home page
